pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kickstarter-ai-scraper"
version = "0.1.0"
description = "Scrape Kickstarter AI/ML project data (Kaggle-equivalent + extras)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "httpx>=0.27",
    "tenacity>=8.2",
    "pydantic>=2.5",
    "pandas>=2.1",
    "pyarrow>=14.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.1",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "respx>=0.21",
    "ruff>=0.3",
]

[project.scripts]
ks-scrape = "kickstarter_scraper.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
